separated add-edit-link-modal into individual components
